+++ DoxygenRule.cmake
@@ -1,4 +1,4 @@
-# Copyright (c) 2012-2014 Stefan Eilemann <eile@eyescale.ch>
+# Copyright (c) 2012-2017 Stefan Eilemann <eile@eyescale.ch>
 
 # Configures a Doxyfile and provides doxygen and doxycopy targets. Relies on
 # TargetHooks installed by Common and must be included after all targets!
@@ -116,12 +116,6 @@ if(NOT DOXYGEN_CONFIG_FILE)
   set(DOXYGEN_CONFIG_FILE ${PROJECT_BINARY_DIR}/doc/Doxyfile)
 endif()
 
-add_custom_target(${PROJECT_NAME}-install
-  ${CMAKE_COMMAND} -P ${PROJECT_BINARY_DIR}/cmake_install.cmake
-  DEPENDS ${PROJECT_NAME}-all)
-set_target_properties(${PROJECT_NAME}-install PROPERTIES
-  EXCLUDE_FROM_DEFAULT_BUILD ON FOLDER ${PROJECT_NAME}/doxygen)
-
 add_custom_target(${PROJECT_NAME}-doxygen
   ${DOXYGEN_EXECUTABLE} ${DOXYGEN_CONFIG_FILE}
   WORKING_DIRECTORY ${PROJECT_BINARY_DIR}/doc
@@ -138,8 +132,7 @@ endif()
 
 if(NOT TARGET doxygen)
   add_custom_target(doxygen)
-  set_target_properties(doxygen PROPERTIES
-    EXCLUDE_FROM_DEFAULT_BUILD ON)
+  set_target_properties(doxygen PROPERTIES EXCLUDE_FROM_DEFAULT_BUILD ON)
 endif()
 add_dependencies(doxygen ${PROJECT_NAME}-doxygen)